+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [21.1.0-beta.2]
+
+### Added
+
+- More copper based items
+  - Copper Shard
+  - Copper Dirty Dust
+  - Copper Crystal
+  - Copper Clump
+- Loot tables for vanilla resources that are produces from modded blocks. I.E Raw copper from Netherack Copper Ore
+
 ## [21.1.0-beta.1]
 
 ### Added